fix: QE 30m preset active 표시 안 되는 버그 수정
qeSetPreset 내 parseInt → parseFloat으로 변경 (parseInt("0.5")=0 으로 비교 실패)

+++ static/js/explorer.js
@@ -106,7 +106,7 @@ async function qeLoadClusters() {
 function qeSetPreset(h) {
   _activeHours = h;
   document.querySelectorAll('.preset-btn').forEach(b => {
-    b.classList.toggle('active', parseInt(b.dataset.h) === h);
+    b.classList.toggle('active', parseFloat(b.dataset.h) === h);
   });
   const now     = new Date();
   const fromStr = _toDatetimeLocal(new Date(now.getTime() - h * 3600 * 1000));
